+++ packages/auth-core/tests/unit/jest.config.ts
@@ -3,10 +3,10 @@ import { jestUnitConfig } from '@ttoss/config';
 export default jestUnitConfig({
   coverageThreshold: {
     global: {
-      statements: 99,
-      branches: 97.9,
+      statements: 99.1,
+      branches: 97.95,
       functions: 100,
-      lines: 99,
+      lines: 99.1,
     },
   },
 });