+++ packages/oauth-client/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.3](https://github.com/ttoss/ttoss/compare/@ttoss/oauth-client@0.2.2...@ttoss/oauth-client@0.2.3) (2026-07-30)
+
+**Note:** Version bump only for package @ttoss/oauth-client
+
 ## [0.2.2](https://github.com/ttoss/ttoss/compare/@ttoss/oauth-client@0.2.1...@ttoss/oauth-client@0.2.2) (2026-07-29)
 
 **Note:** Version bump only for package @ttoss/oauth-client